Two big changes - Move list types into types file - Move list configuraiton into list config service
