+++ src/frontend/app/features/applications/application/application-tabs-base/tabs/events-tab/events-tab.component.html
@@ -1,4 +1,3 @@
 <div class="events-tab">
-  <app-list [text]="{ title: 'Events'}" [enableFilter]="false">
-  </app-list>
+  <app-list></app-list>
 </div>
