+++ src/frontend/app/features/applications/application/application-tabs-base/tabs/instances-tab/instances-tab.component.html
@@ -11,4 +11,4 @@
     </app-tile>
   </app-tile-group>
 </app-tile-grid>
-<app-list [enableFilter]="false"></app-list>
+<app-list></app-list>
