+++ src/frontend/app/features/applications/application/application-tabs-base/tabs/variables-tab/variables-tab.component.html
@@ -1,6 +1,6 @@
 <div class="variables-tab">
   <div class="variables-tab__env-table">
-    <app-list [text]="{ title: 'Environment Variables', filter: 'Filter Variables'}" [enableFilter]="true" [addForm]="addForm" [tableFixedRowHeight]="false">
+    <app-list [addForm]="addForm">
       <div app-table-add>
         <form #addForm="ngForm" novalidate>
           <mat-form-field floatPlaceholder="never">
