+++ src/frontend/app/features/applications/application/application-tabs-base/tabs/variables-tab/variables-tab.component.ts
@@ -2,11 +2,13 @@ import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs/Observable';
 
-import { CfAppEvnVarsDataSource } from '../../../../../../shared/components/list/list-types/app-variables/cf-app-variables-data-source';
+import {
+  CfAppEvnVarsDataSource,
+} from '../../../../../../shared/components/list/list-types/app-variables/cf-app-variables-data-source';
 import {
   CfAppVariablesListConfigService,
 } from '../../../../../../shared/components/list/list-types/app-variables/cf-app-variables-list-config.service';
-import { ListConfig } from '../../../../../../shared/components/list/list.component';
+import { ListConfig } from '../../../../../../shared/components/list/list.component.types';
 import { AppState } from '../../../../../../store/app-state';
 import { ApplicationService } from '../../../../application.service';
