+++ src/frontend/app/features/applications/routes/routes.component.html
@@ -1,6 +1,5 @@
 <div class="routes">
   <div class="routes-table">
-    <app-list [text]="{ title: 'Routes'}" [enableFilter]="false">
-    </app-list>
+    <app-list></app-list>
   </div>
 </div>
