+++ src/frontend/app/features/applications/routes/routes.component.ts
@@ -1,11 +1,13 @@
-import { CfAppRoutesListConfigService } from '../../../shared/components/list/list-types/app-route/cf-app-routes-list-config.service';
-import { ListConfig } from '../../../shared/components/list/list.component';
-import { ApplicationService } from '../application.service';
-import { CfAppRoutesDataSource } from '../../../shared/components/list/list-types/app-route/cf-app-routes-data-source';
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
 import { Store } from '@ngrx/store';
+
+import { CfAppRoutesDataSource } from '../../../shared/components/list/list-types/app-route/cf-app-routes-data-source';
+import {
+  CfAppRoutesListConfigService,
+} from '../../../shared/components/list/list-types/app-route/cf-app-routes-list-config.service';
+import { ListConfig } from '../../../shared/components/list/list.component.types';
 import { AppState } from '../../../store/app-state';
+import { ApplicationService } from '../application.service';
 
 @Component({
   selector: 'app-routes',
