+++ src/frontend/app/features/endpoints/endpoints-page/endpoints-page.component.html
@@ -8,5 +8,4 @@
 </app-page-header>
 
 <app-endpoints-missing [showSnackForNoneConnected]="true"></app-endpoints-missing>
-<app-list [text]="{ filter: 'Filter Endpoints'}" [enableFilter]="true" [tableFixedRowHeight]="true" [hidden]="!(endpointsService.haveRegistered$ | async)">
-</app-list>
+<app-list [hidden]="!(endpointsService.haveRegistered$ | async)"></app-list>
