+++ src/frontend/app/shared/components/list/list-table/table-cell-actions/table-cell-actions.component.ts
@@ -1,9 +1,11 @@
 /* tslint:disable:no-access-missing-member https://github.com/mgechev/codelyzer/issues/191/ */
-import { IListAction, ListConfig } from '../../list.component';
-import { Component, OnInit } from '@angular/core';
-import { TableCellCustom } from '../table-cell/table-cell-custom';
+import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
+
 import { AppState } from '../../../../../store/app-state';
+import { IListAction, ListConfig } from '../../list.component.types';
+import { TableCellCustom } from '../table-cell/table-cell-custom';
+
 
 @Component({
   selector: 'app-table-cell-actions',
