+++ src/frontend/app/shared/components/list/list-types/app/cf-app-config.service.ts
@@ -8,11 +8,12 @@ import { APIResource } from '../../../../../store/types/api.types';
 import { CfOrgSpaceDataService, CfOrgSpaceItem } from '../../../../data-services/cf-org-space-service.service';
 import { ApplicationStateService } from '../../../application-state/application-state.service';
 import { ITableColumn } from '../../list-table/table.types';
-import { IListConfig, IListMultiFilterConfig, ListViewTypes } from '../../list.component';
+import { IListConfig, IListMultiFilterConfig, ListViewTypes } from '../../list.component.types';
 import { CfAppsDataSource } from './cf-apps-data-source';
 import { TableCellAppInstancesComponent } from './table-cell-app-instances/table-cell-app-instances.component';
 import { TableCellAppNameComponent } from './table-cell-app-name/table-cell-app-name.component';
 import { TableCellAppStatusComponent } from './table-cell-app-status/table-cell-app-status.component';
+import { CardAppComponent } from './card/card-app.component';
 
 
 @Injectable()
@@ -62,6 +63,12 @@ export class CfAppConfigService implements IListConfig<APIResource> {
   ];
   pageSizeOptions = [9, 45, 90];
   viewType = ListViewTypes.BOTH;
+  text = {
+    title: '',
+    filter: 'Filter Applications'
+  };
+  enableTextFilter = true;
+  cardComponent = CardAppComponent;
 
   getGlobalActions = () => null;
   getMultiActions = () => null;
