+++ src/frontend/app/shared/components/list/list-types/endpoint/endpoints-list-config.service.ts
@@ -1,5 +1,3 @@
-import { ResetPagination } from '../../../../../store/actions/pagination.actions';
-import { isLowerCase } from 'tslint/lib/utils';
 import { Injectable } from '@angular/core';
 import { MatDialog } from '@angular/material';
 import { Store } from '@ngrx/store';
@@ -8,22 +6,20 @@ import {
   ConnectEndpointDialogComponent,
 } from '../../../../../features/endpoints/connect-endpoint-dialog/connect-endpoint-dialog.component';
 import { DisconnectCnis, UnregisterCnis } from '../../../../../store/actions/cnsis.actions';
-import { RouterNav } from '../../../../../store/actions/router.actions';
+import { ResetPagination } from '../../../../../store/actions/pagination.actions';
 import { ShowSnackBar } from '../../../../../store/actions/snackBar.actions';
 import { GetSystemInfo } from '../../../../../store/actions/system.actions';
 import { AppState } from '../../../../../store/app-state';
 import { CNSISEffect } from '../../../../../store/effects/cnsis.effects';
 import { selectUpdateInfo } from '../../../../../store/selectors/api.selectors';
 import { CNSISModel, cnsisStoreNames } from '../../../../../store/types/cnsis.types';
-import { IGlobalListAction, IListAction, IListConfig, IMultiListAction, ListViewTypes } from '../../list.component';
-import {
-  TableCellEndpointStatusComponent,
-} from './table-cell-endpoint-status/table-cell-endpoint-status.component';
 import { TableCellActionsComponent } from '../../list-table/table-cell-actions/table-cell-actions.component';
 import { TableCellSelectComponent } from '../../list-table/table-cell-select/table-cell-select.component';
 import { TableHeaderSelectComponent } from '../../list-table/table-header-select/table-header-select.component';
 import { ITableColumn } from '../../list-table/table.types';
+import { IListAction, IListConfig, IMultiListAction, ListViewTypes } from '../../list.component.types';
 import { EndpointsDataSource } from './endpoints-data-source';
+import { TableCellEndpointStatusComponent } from './table-cell-endpoint-status/table-cell-endpoint-status.component';
 
 
 function getEndpointTypeString(endpoint: CNSISModel): string {
@@ -148,6 +144,12 @@ export class EndpointsListConfigService implements IListConfig<CNSISModel> {
   dataSource: EndpointsDataSource;
   pageSizeOptions = [9, 45, 90];
   viewType = ListViewTypes.TABLE_ONLY;
+  text = {
+    title: '',
+    filter: 'Filter Endpoints'
+  };
+  enableTextFilter = true;
+  tableFixedRowHeight = true;
 
   private handleAction(item, effectKey, handleChange) {
     const disSub = this.store.select(selectUpdateInfo(
