+++ src/frontend/app/shared/components/list/list.component.html
@@ -4,7 +4,7 @@
       <mat-card-header>
         <div class="listComponent__header__top">
           <div class="listComponent__header__left spacer">
-            <div [hidden]="(dataSource.isSelecting$ | async) || (dataSource.isAdding$ | async)">{{ text?.title }}</div>
+            <div [hidden]="(dataSource.isSelecting$ | async) || (dataSource.isAdding$ | async)">{{ config.text?.title }}</div>
             <div [hidden]="!(dataSource.isSelecting$ | async)">{{dataSource.selectedRows.size}} Selected</div>
           </div>
           <div class="listComponent__header__right">
@@ -37,7 +37,7 @@
                 <mat-icon>add</mat-icon>
               </button>
             </div>
-            <div *ngIf="listConfigService.viewType === 'both' && !(dataSource.isSelecting$ | async) && (dataSource.pagination$ | async)?.totalResults > 0">
+            <div *ngIf="config.viewType === 'both' && !(dataSource.isSelecting$ | async) && (dataSource.pagination$ | async)?.totalResults > 0">
               <button mat-icon-button [disabled]="(dataSource.isLoadingPage$ | async)" (click)="updateListView('cards');" *ngIf="(dataSource.view$ | async)==='table'">
                   <mat-icon>grid_on</mat-icon>
                 </button>
@@ -76,9 +76,9 @@
                 </mat-select>
               </mat-form-field>
             </div>
-            <div class="filter" [hidden]="!enableFilter || ((dataSource.pagination$ | async)?.totalResults === 0 && !filter) || (dataSource.isSelecting$ | async)">
+            <div class="filter" [hidden]="!config.enableTextFilter || ((dataSource.pagination$ | async)?.totalResults === 0 && !filter) || (dataSource.isSelecting$ | async)">
               <mat-form-field floatPlaceholder="never">
-                <input matInput [ngModel]="filterString" #filter="ngModel" [disabled]="(dataSource.isLoadingPage$ | async)" name="filter" placeholder="{{text?.filter || 'Filter'}}">
+                <input matInput [ngModel]="filterString" #filter="ngModel" [disabled]="(dataSource.isLoadingPage$ | async)" name="filter" placeholder="{{config.text?.filter || 'Filter'}}">
               </mat-form-field>
             </div>
           </div>
@@ -92,8 +92,9 @@
     </div>
   </div>
   <div class="listComponent__body">
-    <app-cards *ngIf="(dataSource.view$ | async) === 'cards'" #cards [dataSource]="dataSource" [component]="cardComponent" [hidden]="!(dataSource.isLoadingPage$ | async) && (dataSource.pagination$ | async)?.totalResults === 0"></app-cards>
-    <app-table *ngIf="(dataSource.view$ | async) === 'table'" #table [dataSource]="dataSource" [paginationController]="paginationController" [columns]="columns" [text]="text" [enableFilter]="true" [fixedRowHeight]="tableFixedRowHeight" [hidden]="!(dataSource.isLoadingPage$ | async) && (dataSource.pagination$ | async)?.totalResults === 0">
+    <app-cards *ngIf="(dataSource.view$ | async) === 'cards'" #cards [dataSource]="dataSource" [component]="config.cardComponent" [hidden]="!(dataSource.isLoadingPage$ | async) && (dataSource.pagination$ | async)?.totalResults === 0"></app-cards>
+    <app-table *ngIf="(dataSource.view$ | async) === 'table'" #table [dataSource]="dataSource" [paginationController]="paginationController" [columns]="columns" [text]="config.text" [enableFilter]="config.enableTextFilter" [fixedRowHeight]="config.tableFixedRowHeight"
+      [hidden]="!(dataSource.isLoadingPage$ | async) && (dataSource.pagination$ | async)?.totalResults === 0">
     </app-table>
     <mat-paginator #paginator [hidden]="(dataSource.pagination$ | async)?.totalResults === 0 || (dataSource.pagination$ | async)?.totalResults < (dataSource.pagination$ | async)?.pageSize" [pageSizeOptions]="paginator.pageSizeOptions" [pageSize]="paginator.pageSizeOptions[0]">
     </mat-paginator>
