+++ src/frontend/app/shared/components/list/list.component.ts
@@ -1,4 +1,3 @@
-
 import {
   AfterViewInit,
   ChangeDetectorRef,
@@ -9,93 +8,26 @@ import {
   OnInit,
   Type,
   ViewChild,
-  Injector,
 } from '@angular/core';
 import { NgForm, NgModel } from '@angular/forms';
-import { MatPaginator, MatSelect, PageEvent, SortDirection } from '@angular/material';
+import { MatPaginator, MatSelect, SortDirection } from '@angular/material';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs/Observable';
 import { Subscription } from 'rxjs/Subscription';
 
-import {
-  ListFilter,
-  ListPagination,
-  ListSort,
-  ListView,
-  SetListViewAction,
-} from '../../../store/actions/list.actions';
+import { ListFilter, ListPagination, ListSort, ListView, SetListViewAction } from '../../../store/actions/list.actions';
 import { AppState } from '../../../store/app-state';
 import { IListDataSource } from './data-sources-controllers/list-data-source-types';
+import { IListPaginationController, ListPaginationController } from './data-sources-controllers/list-pagination-controller';
 import { ITableColumn, ITableText } from './list-table/table.types';
-import { StaticInjector } from '@angular/core/src/di/injector';
-import { ListDataSource } from './data-sources-controllers/list-data-source';
-import { ListPaginationController, IListPaginationController } from './data-sources-controllers/list-pagination-controller';
-import { distinctUntilChanged, map, tap } from 'rxjs/operators';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import {
+  IGlobalListAction,
+  IListAction,
+  IListMultiFilterConfig,
+  IMultiListAction,
+  ListConfig,
+} from './list.component.types';
 
-export enum ListViewTypes {
-  CARD_ONLY = 'cardOnly',
-  TABLE_ONLY = 'tableOnly',
-  BOTH = 'both'
-}
-
-export interface IListConfig<T> {
-  getGlobalActions: () => IGlobalListAction<T>[];
-  getMultiActions: () => IMultiListAction<T>[];
-  getSingleActions: () => IListAction<T>[];
-  getColumns: () => ITableColumn<T>[];
-  getDataSource: () => IListDataSource<T>;
-  getMultiFiltersConfigs: () => IListMultiFilterConfig[];
-  isLocal?: boolean;
-  pageSizeOptions: Number[];
-  viewType: ListViewTypes;
-}
-
-export interface IListMultiFilterConfig {
-  key: string;
-  label: string;
-  list$: Observable<IListMultiFilterConfigItem[]>;
-  loading$: Observable<boolean>;
-  select: BehaviorSubject<any>;
-}
-
-export interface IListMultiFilterConfigItem {
-  label: string;
-  item: any;
-  value: string;
-}
-
-export class ListConfig implements IListConfig<any> {
-  isLocal = false;
-  pageSizeOptions = [9, 45, 90];
-  viewType = ListViewTypes.BOTH;
-  getGlobalActions = () => null;
-  getMultiActions = () => null;
-  getSingleActions = () => null;
-  getColumns = () => null;
-  getDataSource = () => null;
-  getMultiFiltersConfigs = () => [];
-}
-
-export interface IBaseListAction<T> {
-  icon: string;
-  label: string;
-  description: string;
-  visible: (row: T) => boolean;
-  enabled: (row: T) => boolean;
-}
-
-export interface IListAction<T> extends IBaseListAction<T> {
-  action: (item: T) => void;
-}
-
-export interface IMultiListAction<T> extends IBaseListAction<T> {
-  action: (items: T[]) => void;
-}
-
-export interface IGlobalListAction<T> extends IBaseListAction<T> {
-  action: () => void;
-}
 
 @Component({
   selector: 'app-list',
@@ -106,10 +38,6 @@ export interface IGlobalListAction<T> extends IBaseListAction<T> {
 export class ListComponent<T> implements OnInit, OnDestroy, AfterViewInit {
   private uberSub: Subscription;
 
-  @Input('text') text = null as ITableText;
-  @Input('enableFilter') enableFilter = false;
-  @Input('tableFixedRowHeight') tableFixedRowHeight = false;
-  @Input('cardComponent') cardComponent: Type<{}>;
   @Input('addForm') addForm: NgForm;
 
   @ViewChild(MatPaginator) paginator: MatPaginator;
@@ -141,21 +69,21 @@ export class ListComponent<T> implements OnInit, OnDestroy, AfterViewInit {
   constructor(
     private _store: Store<AppState>,
     private cd: ChangeDetectorRef,
-    private listConfigService: ListConfig
+    public config: ListConfig
   ) { }
 
   ngOnInit() {
 
-    this.globalActions = this.listConfigService.getGlobalActions();
-    this.multiActions = this.listConfigService.getMultiActions();
-    this.singleActions = this.listConfigService.getSingleActions();
-    this.columns = this.listConfigService.getColumns();
-    this.dataSource = this.listConfigService.getDataSource();
-    this.multiFilterConfigs = this.listConfigService.getMultiFiltersConfigs();
+    this.globalActions = this.config.getGlobalActions();
+    this.multiActions = this.config.getMultiActions();
+    this.singleActions = this.config.getSingleActions();
+    this.columns = this.config.getColumns();
+    this.dataSource = this.config.getDataSource();
+    this.multiFilterConfigs = this.config.getMultiFiltersConfigs();
 
     this.paginationController = new ListPaginationController(this._store, this.dataSource);
 
-    this.paginator.pageSizeOptions = this.listConfigService.pageSizeOptions;
+    this.paginator.pageSizeOptions = this.config.pageSizeOptions;
     const paginationStoreToWidget = this.paginationController.pagination$.do((pagination: ListPagination) => {
       this.paginator.length = pagination.totalResults;
       this.paginator.pageIndex = pagination.pageIndex - 1;
